# apps/api-service/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MCPFactory API Service",
    "description": "API Gateway for MCPFactory. Handles authentication, proxies to internal services, and exposes the public REST API.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.mcpfactory.org"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Bearer JWT from Clerk (dashboard)"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for MCP clients and service-to-service communication"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": [
          "error"
        ]
      },
      "WebhookResponse": {
        "type": "object",
        "properties": {
          "received": {
            "type": "boolean"
          }
        },
        "required": [
          "received"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/webhooks/clerk": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Clerk webhook receiver",
        "description": "Receives Clerk lifecycle events (user.created, session.created). Verified via svix signature headers.",
        "responses": {
          "200": {
            "description": "Webhook received",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid signature or missing headers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Webhook not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}